% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_methods.R
\name{dbExecute}
\alias{dbExecute}
\title{Execute any SQL statement from Nuvolos.cloud}
\usage{
dbExecute(sql, con)
}
\arguments{
\item{sql}{SQL statement to be executed. Note that quoting the tables is needed only if the table name is case sensitive (it contains both upper and lowercase letters or special chars).}

\item{con}{pyodbc connection object when using on Nuvolos and an sqlalchemy connection object when using from a local device. The object can be created by get_connection function.}
}
\value{
Returns the result of python's execute method.
}
\description{
Function dbExecute(sql, con).
Executes any SQL statement in the connected Nuvolos schema.
}
\examples{
dbExecute("DROP TABLE table", con)
dbExecute("DROP TABLE \"Table\"", con)
}
