% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_methods.R
\name{execute}
\alias{execute}
\title{Execute any SQL statement from Nuvolos.cloud}
\usage{
execute(sql, dbname = NULL, schemaname = NULL)
}
\arguments{
\item{sql}{SQL statement to be executed. Note that quoting the tables is needed only if the table name is case sensitive (it contains both upper and lowercase letters or special chars).}

\item{dbname}{The name of the database from/in which the statement will be executed.}

\item{schemaname}{The name of the schema from/in which the statement will be executed.}
}
\value{
Returns the result of python's execute method.
}
\description{
Function execute(sql, dbname, schemaname).
Executes any SQL statement in the connected Nuvolos schema.
On Nuvolos the database and schema are by default the ones the user is working in, from local machine they need to be specified.
}
\examples{
execute("DROP TABLE table")
execute("DROP TABLE table", dbname = "space_1", schemaname = "test_schema")
}
