% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataHubCon.R
\name{getDataHubCon}
\alias{getDataHubCon}
\title{R connector for DataHub.ac}
\usage{
getDataHubCon(dbname, schemaname)
}
\arguments{
\item{dbname}{The database (organization + space) to connect to}

\item{schemaname}{The schema (instance + state) to connect to}
}
\description{
Function getDataHubCon(dbname, schemaname)
Creates a new connection to schema dbname.schemaname
}
