% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connection.R
\name{input_nuvolos_credential}
\alias{input_nuvolos_credential}
\title{Function input_nuvolos_credential()
Using outside of Nuvolos only. This helps the user to store credentials safely at local device.}
\usage{
input_nuvolos_credential()
}
\description{
Function input_nuvolos_credential()
Using outside of Nuvolos only. This helps the user to store credentials safely at local device.
}
