% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_methods.R
\name{read_sql}
\alias{read_sql}
\title{Execute SQL SELECT statements from Nuvolos.cloud}
\usage{
read_sql(sql, dbname = NULL, schemaname = NULL, index_col = NULL)
}
\arguments{
\item{sql}{SQL statement to be executed. Note that quoting the tables is needed only if the table name is case sensitive (it contains both upper and lowercase letters or special chars).}

\item{dbname}{The name of the database from which the SELECT statement will be executed.}

\item{schemaname}{The name of the schema from which the SELECT statement will be executed.}

\item{index_col}{Name of the column to set as index. Default is NULL.}
}
\value{
Returns an R dataframe object. When date format columns are in the table, they are returned as Date.
}
\description{
Function read_sql(sql, dbname, schemaname, parse_dates).
Executes SELECT SQL statements in the connected Nuvolos schema.
On Nuvolos the database and schema are by default the ones the user is working in, from local machine they need to be specified.
}
\examples{
db <- read_sql("SELECT * FROM table")
db <- read_sql("SELECT * FROM table", dbname = "space_1", schemaname = "test_schema", index_col = "index")
}
