% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_methods.R
\name{dbGetQuery}
\alias{dbGetQuery}
\title{Execute SQL SELECT statements from Nuvolos.cloud}
\usage{
dbGetQuery(sql, con)
}
\arguments{
\item{sql}{SQL statement to be executed. Make sure to use quotes around table names.}

\item{con}{SQLalchemy connection object that can be created by get_connection function.}
}
\value{
Returns an R dataframe object.
}
\description{
Function dbGetQuery(sql, con)
Executes SELECT SQL statements in the connected Nuvolos schema.
}
