% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connection.R
\name{get_connection}
\alias{get_connection}
\title{R connector for Nuvolos.cloud}
\usage{
get_connection(dbname, schemaname)
}
\arguments{
\item{dbname}{The database (organization + space) to connect to}

\item{schemaname}{The schema (instance + state) to connect to}
}
\description{
Function get_connection(dbname, schemaname)
Creates a new connection to schema dbname.schemaname. If both arguments are left NULL (empty), get_connection will try to figure out the appropriate parameters by calling get_nuvolos_db_path().
}
