% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connection.R
\name{get_nuvolos_db_path}
\alias{get_nuvolos_db_path}
\title{Function get_nuvolos_db_path()
In Nuvolos applications returns the database name and schema name.
Outside Nuvolos it will return with an error: please refer to the connection guide for detailed information on database and schema names for out-of-Nuvolos usage.}
\usage{
get_nuvolos_db_path()
}
\description{
Function get_nuvolos_db_path()
In Nuvolos applications returns the database name and schema name.
Outside Nuvolos it will return with an error: please refer to the connection guide for detailed information on database and schema names for out-of-Nuvolos usage.
}
