% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataHubCon.R
\name{get_nuvolos_db_path}
\alias{get_nuvolos_db_path}
\title{Function get_nuvolos_db_path()
In AAP applications returns the database name and schema name}
\usage{
get_nuvolos_db_path()
}
\description{
Function get_nuvolos_db_path()
In AAP applications returns the database name and schema name
}
