% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataHubCon.R
\name{getDBPath}
\alias{getDBPath}
\title{Function getDBPath()
In AAP applications returns the database name and schema name}
\usage{
getDBPath()
}
\description{
Function getDBPath()
In AAP applications returns the database name and schema name
}
