% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connection.R
\name{input_nuvolos_credential}
\alias{input_nuvolos_credential}
\title{Input credentials to connect Nuvolos.cloud from local device}
\usage{
input_nuvolos_credential()
}
\value{
Returns a pop-up window to enter the credentials.
}
\description{
Function input_nuvolos_credential()
Using outside of Nuvolos only. This helps the user to store credentials safely at local device.
Should be called only when there was a wrong input in the first place. It overwrites the previously entered credentials.
}
