% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_methods.R
\name{read_sql}
\alias{read_sql}
\title{Execute SQL SELECT statements from Nuvolos.cloud}
\usage{
read_sql(sql, dbname = NULL, schemaname = NULL)
}
\arguments{
\item{sql}{SQL statement to be executed. Make sure to use quotes around table names.}

\item{dbname}{The name of the database from which the SELECT statement will be executed.}

\item{schemaname}{The name of the schema from which the SELECT statement will be executed.}
}
\value{
Returns an R dataframe object.
}
\description{
Function read_sql(sql, dbname, schemaname).
Executes SELECT SQL statements in the connected Nuvolos schema.
On Nuvolos the database and schema are by default the ones the user is working in, from local machine they need to be specified.
}
\examples{
db <- read_sql('SELECT * FROM "table"')
db <- read_sql('SELECT * FROM "table"', dbname = "space_1", schemaname = "test_schema")

}
